test(verification): merge hooks tests into SinchEvents
